+++ R/explain_scikitlearn.R
@@ -80,31 +80,28 @@
 #'
 #'
 #' @examples
-
-#' reticulate::use_condaenv("myenv")
-#' have_sklearn <- reticulate::py_module_available("sklearn.ensemble")
 #' library("DALEXtra")
 #' library("DALEX")
-#' if("myenv" %in% reticulate::conda_list()$name){
-#' if(have_sklearn) {
+#' is_conda <- function(){
+#' is_conda <- try(reticulate::conda_binary())
+#' class(is_conda) != "try-error"
+#' }
+#' if(is_conda()) {
 #'    # Explainer build (Keep in mind that 18th column is target)
 #'    titanic_test <- read.csv(system.file("extdata", "titanic_test.csv", package = "DALEXtra"))
 #'    # Keep in mind that when pickle is being built and loaded,
 #'    # not only Python version but libraries versions has to match aswell
 #'    explainer <- explain_scikitlearn(system.file("extdata", "scikitlearn.pkl", package = "DALEXtra"),
-#'    condaenv = "myenv", data = titanic_test[,1:17], y = titanic_test$survived)
-#'    print(model_performance(explainer))
+#'    yml = system.file("extdata", "scikitlearn_unix.yml", package = "DALEXtra"),
+#'    data = titanic_test[,1:17], y = titanic_test$survived)
+#'    plot(model_performance(explainer))
 #'
 #'    # Predictions with newdata
 #'    explainer$model$predict_function(explainer$model, titanic_test[1:10,1:17])
 #'
 #' } else {
-#'   print('Python testing environment is required.')
+#'   print('Conda is required.')
 #' }
-#'} else{
-#'   print("myenv does not exists, use yml argument")
-#'
-#'}
 #'
 #' @rdname explain_scikitlearn
 #' @export
@@ -124,10 +121,6 @@ explain_scikitlearn <-
            precalculate = TRUE) {
     if (!is.null(condaenv) & !is.null(env)) {
       stop("Only one argument from condaenv and env can be different from NULL")
-    }
-    if (!is.null(yml) &
-        is.null(condaenv) & .Platform$OS.type == "unix") {
-      stop("You have to provide condaenv parameter with yml when using unix-like OS")
     }
 
     if (!is.null(yml)) {
@@ -151,7 +144,7 @@ explain_scikitlearn <-
         error = function(e) {
           warning(e, call. = FALSE)
           stop(
-            "reticulate is unable to set new environment due to already using other python.exe, please restart R session. See warnings() for original error",
+            "reticulate is unable to set new environment. Specified envirnonment does not exists or connection cannot be established due to already using other python.exe, please install environment or restart R session. See warnings() for original error",
             call. = FALSE
           )
         }
@@ -167,7 +160,7 @@ explain_scikitlearn <-
         error = function(e) {
           warning(e, call. = FALSE)
           stop(
-            "reticulate is unable to set new environment due to already using other python.exe, please restart R session. See warnings() for original error",
+            "reticulate is unable to set new environment. Specified envirnonment does not exists or connection cannot be established due to already using other python.exe, please install environment or restart R session. See warnings() for original error",
             call. = FALSE
           )
         }