+++ tests/testthat/test_create_env.R
@@ -3,17 +3,34 @@ context("create_env")
 source("objects_for_tests.R")
 
 test_that("creating env", {
-  skip_if_windows()
   skip_if_no_conda()
-  if("myenv" %in% reticulate::conda_list()$name){
+  if ("myenv" %in% reticulate::conda_list()$name) {
     reticulate::conda_remove("myenv")
-    name <- create_env(system.file("extdata", "scikitlearn_unix.yml", package = "DALEXtra"),
-                       condaenv = paste(sub('[/][^/]+$', '', reticulate::conda_binary()), "/..", sep = ""))
+    name <-
+      create_env(
+        system.file("extdata", "scikitlearn_unix.yml", package = "DALEXtra")
+        )
+
   } else{
-      name <- create_env(system.file("extdata", "scikitlearn_unix.yml", package = "DALEXtra"),
-                 condaenv = paste(sub('[/][^/]+$', '', reticulate::conda_binary()), "/..", sep = ""))
+    name <-
+      create_env(
+        system.file("extdata", "scikitlearn_unix.yml", package = "DALEXtra"),
+      )
   }
   expect_is(name, "character")
 
 
 })
+test_that("if check", {
+  if (.Platform$OS.type == "unix") {
+    expect_success(expect_message(create_env(yml = system.file("extdata", "scikitlearn_unix.yml", package = "DALEXtra")),
+                                  "not specified"))
+  }
+  if ("myenv" %in% reticulate::conda_list()$name){
+    expect_success(expect_message(create_env(yml = system.file("extdata", "scikitlearn_unix.yml", package = "DALEXtra")),
+                                  "exists"))
+  }
+})
+
+
+